+++ core/src/main/kotlin/com/rudderstack/sdk/kotlin/core/internals/policies/FlushPoliciesFacade.kt
@@ -1,6 +1,7 @@
 package com.rudderstack.sdk.kotlin.core.internals.policies
 
 import com.rudderstack.sdk.kotlin.core.Analytics
+import com.rudderstack.sdk.kotlin.core.internals.logger.Logger
 
 /**
  * FlushPoliciesFacade manages a collection of different flush policies
@@ -9,14 +10,24 @@ import com.rudderstack.sdk.kotlin.core.Analytics
  * ensuring the appropriate actions are taken based on each policy's state.
  *
  * @property flushPolicies A list of flush policies that this facade will manage.
+ * @property logger The logger instance for logging flush policy actions and states.
  */
-internal class FlushPoliciesFacade(private val flushPolicies: List<FlushPolicy>) {
+internal class FlushPoliciesFacade(
+    private val flushPolicies: List<FlushPolicy>,
+    private val logger: Logger,
+) {
 
-    internal fun shouldFlush(): Boolean = flushPolicies.any {
-        when (it) {
-            is CountFlushPolicy -> it.shouldFlush()
-            is StartupFlushPolicy -> it.shouldFlush()
-            else -> false
+    internal fun shouldFlush(): Boolean {
+        return flushPolicies.any { policy ->
+            val shouldFlush = when (policy) {
+                is CountFlushPolicy -> policy.shouldFlush()
+                is StartupFlushPolicy -> policy.shouldFlush()
+                else -> false
+            }
+            if (shouldFlush) {
+                logger.verbose("FlushPoliciesFacade: Flush triggered by ${policy::class.simpleName}")
+            }
+            shouldFlush
         }
     }