+++ core/src/test/kotlin/com/rudderstack/sdk/kotlin/core/internals/utils/AnalyticsUtilsTest.kt
@@ -18,12 +18,10 @@ import io.mockk.every
 import io.mockk.impl.annotations.MockK
 import io.mockk.mockk
 import io.mockk.mockkStatic
-import io.mockk.spyk
 import io.mockk.verify
 import kotlinx.coroutines.SupervisorJob
 import kotlinx.coroutines.test.StandardTestDispatcher
 import kotlinx.coroutines.test.TestScope
-import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.jupiter.api.Assertions.assertFalse
 import org.junit.jupiter.api.Assertions.assertTrue
 import org.junit.jupiter.api.BeforeEach
@@ -112,7 +110,7 @@ class AnalyticsUtilsTest {
 
         assertFalse(result)
         verify(exactly = 1) {
-            mockLogger.error("Analytics instance has been shutdown. No further operations are allowed.")
+            mockLogger.warn("AnalyticsUtils: Analytics instance has been shutdown. No further operations are allowed")
         }
     }
 
@@ -159,7 +157,7 @@ class AnalyticsUtilsTest {
 
         assertTrue(result)
         verify(exactly = 0) {
-            mockLogger.error("Source is disabled. This operation is not allowed.")
+            mockLogger.warn("Source is disabled. This operation is not allowed")
         }
     }
 
@@ -172,7 +170,7 @@ class AnalyticsUtilsTest {
 
         assertTrue(result) // Server always returns true
         verify(exactly = 0) {
-            mockLogger.error("Source is disabled. This operation is not allowed.")
+            mockLogger.warn("AnalyticsUtils: Source is disabled. This operation is not allowed")
         }
     }
 
@@ -185,7 +183,7 @@ class AnalyticsUtilsTest {
 
         assertFalse(result)
         verify(exactly = 1) {
-            mockLogger.error("Source is disabled. This operation is not allowed.")
+            mockLogger.warn("AnalyticsUtils: Source is disabled. This operation is not allowed")
         }
     }